updated controllers

+++ backend/src/controllers/publications/book.controller.ts
@@ -3,11 +3,14 @@ import { Request, Response } from "express";
 import { AppError } from "../../utils/error";
 import prisma from "../../config/db";
 import response from "../../utils/response";
+import { signedUrl } from "../../s3";
 
 /**
  * @description creating a book
  * @route POST /api/book/create
  * @access Private
+ * @param req
+ * @param res
  */
 export const createBook = asyncHandler(async (req: Request, res: Response) => {
 	const { title, authors, publisher, scopus, doi, isbn, year } = req.body;
@@ -17,7 +20,29 @@ export const createBook = asyncHandler(async (req: Request, res: Response) => {
 	}
 
 	const book = await prisma.book.create({
-		data: { title, authors, publisher, scopus, doi, isbn, year },
+		data: {
+			title,
+			authors: {
+				connect: authors.map((authorId: number) => ({ id: authorId })),
+			},
+			publisher,
+			scopus,
+			doi,
+			isbn,
+			year,
+		},
+		include: {
+			authors: {
+				select: {
+					id: true,
+					name: true,
+					email: true,
+					role: true,
+					profileImage: true,
+					designation: true,
+				}
+			}
+		}
 	});
 
 	response(res, 201, "Book created successfully", { book });
@@ -27,9 +52,34 @@ export const createBook = asyncHandler(async (req: Request, res: Response) => {
  * @description fetching all books
  * @route GET /api/book/get-all
  * @access Public
+ * @param req
+ * @param res
  */
 export const getAllBooks = asyncHandler(async (req: Request, res: Response) => {
-	const books = await prisma.book.findMany();
+	const books = await prisma.book.findMany({
+		orderBy: {
+			createdAt: "desc",
+		},
+		include: {
+			authors: {
+				select: {
+					id: true,
+					name: true,
+					email: true,
+					role: true,
+					profileImage: true,
+					designation: true,
+				}
+			}
+		}
+	});
+
+	for(let book of books) {
+		for(let author of book.authors) {
+			author.profileImage = author.profileImage ? await signedUrl(author.profileImage, 3) : "";
+		}
+	}
+
 	response(res, 200, "Books fetched successfully", { books });
 });
 
@@ -37,17 +87,37 @@ export const getAllBooks = asyncHandler(async (req: Request, res: Response) => {
  * @description fetching a single book by ID
  * @route GET /api/book/:id
  * @access Public
+ * @param req
+ * @param res
  */
 export const getBookById = asyncHandler(async (req: Request, res: Response) => {
-	const { id } = req.params;
+	const { id } = req.query;
+	if (!id) throw new AppError("Please provide id of book", 400);
+
 	const book = await prisma.book.findUnique({
 		where: { id: Number(id) },
+		include: {
+			authors: {
+				select: {
+					id: true,
+					name: true,
+					email: true,
+					role: true,
+					profileImage: true,
+					designation: true,
+				}
+			}
+		}
 	});
 
 	if (!book) {
 		throw new AppError("Book not found", 404);
 	}
-
+	if (book.authors) {
+		for (let author of book.authors) {
+			author.profileImage = author.profileImage ? await signedUrl(author.profileImage, 3) : "";
+		}
+	}
 	response(res, 200, "Book fetched successfully", { book });
 });
 
@@ -55,16 +125,47 @@ export const getBookById = asyncHandler(async (req: Request, res: Response) => {
  * @description updating a book by ID
  * @route PUT /api/book/:id
  * @access Private
+ * @param req
+ * @param res
  */
 export const updateBook = asyncHandler(async (req: Request, res: Response) => {
-	const { id } = req.params;
-	const { title, authors, publisher, scopus, doi, isbn, year } = req.body;
+	const { id, title, authors, publisher, scopus, doi, isbn, year } = req.body;
+
+	if (!id || !title || !authors || !year) {
+		throw new AppError("Please provide id, title, authors, and year", 400);
+	}
 
 	const book = await prisma.book.update({
 		where: { id: Number(id) },
-		data: { title, authors, publisher, scopus, doi, isbn, year },
+		data: {
+			title,
+			authors: {
+				connect: authors.map((authorId: number) => ({ id: authorId })),
+			},
+			publisher,
+			scopus,
+			doi,
+			isbn,
+			year,
+		},
+		include: {
+			authors: {
+				select: {
+					id: true,
+					name: true,
+					email: true,
+					role: true,
+					profileImage: true,
+					designation: true,
+				}
+			}
+		}
 	});
-
+	if (book.authors) {
+		for (let author of book.authors) {
+			author.profileImage = author.profileImage ? await signedUrl(author.profileImage, 3) : "";
+		}
+	}
 	response(res, 200, "Book updated successfully", { book });
 });
 
@@ -72,9 +173,13 @@ export const updateBook = asyncHandler(async (req: Request, res: Response) => {
  * @description deleting a book by ID
  * @route DELETE /api/book/:id
  * @access Private
+ * @param req
+ * @param res
  */
 export const deleteBook = asyncHandler(async (req: Request, res: Response) => {
 	const { id } = req.params;
+	if (!id) throw new AppError("Please provide id of book", 400);
+
 	const book = await prisma.book.delete({
 		where: { id: Number(id) },
 	});
@@ -86,25 +191,42 @@ export const deleteBook = asyncHandler(async (req: Request, res: Response) => {
  * @description fetching all books by user ID
  * @route POST /api/book/get-all-by-user-id
  * @access Private
+ * @param req
+ * @param res
  */
-export const getAllBooksByUserId = asyncHandler(
-	async (req: Request, res: Response) => {
-		const { userId } = req.body;
-		const books = await prisma.book.findMany({
-			where: {
-				authors: {
-					some: {
-						id: userId,
-					},
-				},
-			},
-			orderBy: { createdAt: "desc" },
-		});
+export const getAllBooksByUserId = asyncHandler(async (req: Request, res: Response) => {
+	const userId = req.user?.userId;
 
-		if (!books || books.length === 0) {
-			throw new AppError("No books found for this user", 404);
+	const books = await prisma.book.findMany({
+		where: {
+			authors: {
+				some: {
+					id: userId as number,
+				}
+			}
+		},
+		orderBy: { createdAt: "desc" },
+		include: {
+			authors: {
+				select: {
+					id: true,
+					name: true,
+					email: true,
+					role: true,
+					profileImage: true,
+					designation: true,
+				}
+			}
 		}
+	});
 
-		response(res, 200, "Books fetched successfully", { books });
+	if (!books || books.length === 0) {
+		throw new AppError("No books found for this user", 404);
 	}
-);
+	for (let book of books) {
+		for (let author of book.authors) {
+			author.profileImage = author.profileImage ? await signedUrl(author.profileImage, 3) : "";
+		}
+	}
+	response(res, 200, "Books fetched successfully", { books });
+});
